Worked on the Indiviual Page code
Changed itemHTML such that it actually uses my airtable terms.
Changed logo on index to have a break before and after so that it isn't pressed up right against the top of the page and the recipies.
Interesting Idea Ivan gave me:
Have it so that the recipies automatically fill the entire column.